Move bootstrap to match pingora docs
Not entirely sure if this actually changes the behaviour of pingora or
not. The docs are not entirely clear whether it will try to take over
connections already, or only when the server is run using run_forever. I
assume the latter since that's when the main loop actually kicks in..